rat-project/requirements.txt
cryptography>=3.4.8
Pillow>=8.3.0
psutil>=5.8.0
requests>=2.25.1
pyinstaller>=4.5.1